Refactor admin views and routing; reorganize imports, enhance AdminBar links, and add new pages for managing orders, users, menu, reservations, reviews, and restaurant information.
