+++ src/lang/fi.json
@@ -137,5 +137,6 @@
   "admin_page.editmenu_page_link": "Muokkaa ruokalistaa",
   "admin_page.editabout_page_link": "Muokkaa tietoja",
   "admin_page.reviews_page_link": "Arvostelut",
-  "admin_page.reservations_page_link": "Pöytävaraukset"
+  "admin_page.reservations_page_link": "Pöytävaraukset",
+  "admin_page.description": "Ylläpitosivulla voit hallita ravintolan toimintoja. Voit tarkastella ja hallita tilauksia, käyttäjiä ruokalistaa, arvosteluja ja pöytävarauksia. Voit myös muokata ravintolan tietoja."
 }
