+++ src/views/admin/Orders.jsx
@@ -1,0 +1,13 @@
+const Orders = () => {
+  return (
+    <div>
+      <h1>Manage Orders</h1>
+      <p>
+        This page is for managing orders: view orders, edit order status, delete
+        orders.
+      </p>
+    </div>
+  );
+};
+
+export default Orders;
