+++ src/views/admin/Reviews.jsx
@@ -2,7 +2,7 @@ const Reviews = () => {
   return (
     <div>
       <h1>Manage Reviews</h1>
-      <p>This page is for manging restaurant reviews.</p>
+      <p>This page is for viewing restaurant reviews.</p>
     </div>
   );
 };
