+++ src/views/admin/Users.jsx
@@ -2,7 +2,7 @@ const Users = () => {
   return (
     <div>
       <h1>Manage Users</h1>
-      <p>This page is for manging users.</p>
+      <p>This page is for manging users: users list, edit user details.</p>
     </div>
   );
 };
