--- src/views/adminviews/Orders.jsx
@@ -1,10 +1,0 @@
-const Orders = () => {
-  return (
-    <div>
-      <h1>Manage Orders</h1>
-      <p>This page is for managing orders.</p>
-    </div>
-  );
-};
-
-export default Orders;
